A couple of minor bug fixes
 - Correct pocket binding's position
 - Ensure pocket binding is not wrapped on the network
 - Fix NPE in network overlay

+++ src/main/java/org/squiddev/cctweaks/core/utils/EntityPosition.java
@@ -32,7 +32,7 @@ public class EntityPosition implements IWorldPosition {
 			y = (int) entity.posY;
 		}
 
-		return new BlockPos((int) entity.posX, y, (int) entity.posZ);
+		return new BlockPos(entity.posX, y, entity.posZ);
 	}
 
 }
